Update README.md
When I have HPC access, I'll work on transferring the data files we have on this array to the webserver so they can be included on this page. 

I'm not sure how much we have (compared to mouse), but I wanted to include the same sections as Mouse for the time being for sake of continuity.